Correct statement of Apéry's theorem.
The previous statement was not what we expect (!). It was stating that the
sequence of partial sums defining zeta(3) was not a constant sequence.
Which is true but obvious. What we want is the fact that it cannot be in the
Cauchy equivalence class of any constant sequence of rational numbers.
But fortunatly the proof script was proving the latter correct formal sentence.
So we only need to change one character, to replace Leibniz equality by Cauchy
 equivalence, in the CR scope :-).