Reviewer's trick to decrease the size of explicit auxiliary constants.
The proof of lemma 6.8 requires bounding a product and a naive proof
involves large numbers in the course of the calculation. This is a trick
to control the number of digits involved in calculations (<= 20).

Unfortunately, this relies in turn on the evaluation of binomial coefficients,
and there is no support to automate the translation from type nat to Z.
The script is hence large and made of ugly copy-pasting. To be improved.

Side note: installing manifest-positive based automation has nasty side effects
by slowing down some //, to the point of seemingly non-terminating
tactic lines. I am considering reverting.